+++ layrz_sdk/entities/message.py
@@ -1,15 +1,13 @@
 """Message entity"""
 
 from datetime import datetime
-from typing import Any, TypeAlias
+from typing import Any
 
-from pydantic import BaseModel, Field
+from pydantic import BaseModel, Field, field_validator
 
 from layrz_sdk.constants import UTC
-
-from .position import Position
-
-PayloadType: TypeAlias = dict[str, Any]
+from layrz_sdk.entities.geofence import Geofence
+from layrz_sdk.entities.position import Position
 
 
 class Message(BaseModel):
@@ -27,11 +25,11 @@ class Message(BaseModel):
     default_factory=lambda: Position(),
     description='Current position of the device',
   )
-  payload: PayloadType = Field(
+  payload: dict[str, Any] = Field(
     default_factory=dict,
     description='Payload data of the device message',
   )
-  sensors: PayloadType = Field(
+  sensors: dict[str, Any] = Field(
     default_factory=dict,
     description='Sensor data of the device message',
   )
@@ -39,3 +37,19 @@ class Message(BaseModel):
     default_factory=lambda: datetime.now(UTC),
     description='Timestamp when the message was received',
   )
+
+  geofences: list[Geofence] = Field(
+    default_factory=list,
+    description='List of geofences associated with the message',
+  )
+
+  @field_validator('geofences', mode='before')
+  def _validate_geofences(cls, value: Any) -> list[Geofence]:
+    """Validate geofences"""
+    if value is None:
+      return []
+
+    if not isinstance(value, list):
+      return []
+
+    return value